fix(AxisLabels): update labels when data range changes
